:hammer: [chore] Font, Color, UIScreen, Fox 위치 수정 #70
