+++ Yeowoo/Sources/Global/Color+.swift
@@ -1,0 +1,29 @@
+//
+//  Color+.swift
+//  Yeowoo
+//
+//  Created by Kim SungHun on 2023/08/02.
+//
+
+import SwiftUI
+
+extension Color{
+	static let mainColor = Color("mainColor")
+	static let backgroundGray = Color("backgroundGray")
+	static let warningRed = Color("warningRed")
+	static let noGray = Color(red: 100 / 255, green: 100 / 255, blue: 100 / 255)
+	static let alarmGray = Color(red: 150 / 255, green: 150 / 255, blue: 150 / 255)
+	static let whiteGray = Color(red: 249 / 255, green: 249 / 255, blue: 249 / 255)
+	static let whiteGray2 = Color(red: 243 / 255, green: 243 / 255, blue: 243 / 255)
+	static let clicked = Color(red: 238 / 255, green: 252 / 255, blue: 254 / 255)
+	static let unclicked = Color(red: 171 / 255, green: 241 / 255, blue: 248 / 255)
+	static let friendGray = Color(red: 47 / 255, green: 47 / 255, blue: 47 / 255)
+	static let circleGray = Color(red: 214 / 255, green: 214 / 255, blue: 214 / 255)
+	
+	static let sceneColor = Color(red: 223 / 255, green: 250 / 255, blue: 81 / 255, opacity: 0.6)
+	static let captaionColor = Color(red: 255 / 255, green: 99 / 255, blue: 13 / 255, opacity: 0.4)
+	static let paparazziColor = Color(red: 88 / 255, green: 228 / 255, blue: 241 / 255, opacity: 0.4)
+	static let activityColor = Color(red: 255 / 255, green: 52 / 255, blue: 100 / 255, opacity: 0.4)
+	static let eatColor = Color(red: 255 / 255, green: 227 / 255, blue: 83 / 255, opacity: 0.6)
+	static let normalColor = Color(red: 25 / 255, green: 25 / 255, blue: 25 / 255, opacity: 0.4)
+}
