+++ Yeowoo/Sources/Global/Font+.swift
@@ -115,4 +115,18 @@ extension Font {
         }
         return size
     }
+	
+	static let semiContent = Font.system(size: 20, weight: .bold, design: .default)
+}
+
+struct SubTitleFont: ViewModifier {
+	func body(content: Content) -> some View {
+		content
+			.font(.title2)
+			.foregroundColor(.black)
+			.fontWeight(.bold)
+			.frame(maxWidth: .infinity, alignment: .leading)
+			.padding(.leading, 20)
+			.padding(.top, 20)
+	}
 }
