+++ Yeowoo/Sources/Models/Fox.swift
